@@ -1,4 +1,5 @@
-﻿using AppBroker.Core.Javascript;
+﻿using AppBroker.Core.Configuration;
+using AppBroker.Core.Javascript;
 using AppBroker.Core.Managers;
 
 namespace AppBroker.Core;
@@ -12,6 +13,7 @@ public interface IInstanceContainer
     IHistoryManager HistoryManager { get; }
     IDeviceStateManager DeviceStateManager { get; }
     JavaScriptEngineManager JavaScriptEngineManager { get; }
+    IConfigManager ConfigManager { get; }
 
     T GetDynamic<T>() where T : class;
     void RegisterDynamic<T>(T instance) where T : class;
